add problem 2942

+++ src/com/liucongblog/leetcode/editor/cn/FindWordsContainingCharacter_2942.java
@@ -58,7 +58,7 @@ class Solution {
     public List<Integer> findWordsContaining(String[] words, char x) {
         List<Integer> list = new ArrayList<>();
         for (int i = 0; i < words.length; i++) {
-            if (words[i].indexOf((int)x)>=0) {
+            if (words[i].indexOf( x)>=0) {
                 list.add(i);
             }
         }
